--- src/main/java/racingcar/domain/car/dto/response/CarRacingDto.java
@@ -1,7 +1,0 @@
-package racingcar.domain.car.dto.response;
-
-public record CarRacingDto(
-        String carName,
-        int coordinate
-) {
-}
