--- src/main/java/racingcar/domain/car/dto/response/CarsRacingDto.java
@@ -1,7 +1,0 @@
-package racingcar.domain.car.dto.response;
-
-import java.util.List;
-
-public record CarsRacingDto(List<CarRacingDto> carRacing) {
-    
-}
