+++ src/main/java/racingcar/domain/game/RoundResult.java
@@ -5,4 +5,7 @@ import java.util.List;
 public record RoundResult(
         List<MoveResult> moveResults
 ) {
+    public MoveResult getAt(final int index) {
+        return moveResults.get(index);
+    }
 }
