+++ src/main/java/racingcar/view/OutputView.java
@@ -1,9 +1,9 @@
 package racingcar.view;
 
 import racingcar.component.ComponentWriter;
-import racingcar.component.out.GameResultComponent;
-import racingcar.component.out.RoundResultComponent;
-import racingcar.domain.car.dto.response.CarsRacingDto;
+import racingcar.component.output.GameResultComponent;
+import racingcar.component.output.RoundResultComponent;
+import racingcar.domain.car.dto.output.CarsRacingDto;
 import racingcar.domain.game.CarsFinishedDto;
 
 public final class OutputView {
