+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.dto.response.CarRacingDto;
+import racingcar.domain.car.dto.output.CarRacingDto;
 import racingcar.domain.move.MoveCommand;
 
 final class CarTest {
