+++ src/test/java/racingcar/domain/car/CarsTest.java
@@ -6,9 +6,9 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.dto.request.CreateCarsRacing;
-import racingcar.domain.car.dto.response.CarRacingDto;
-import racingcar.domain.car.dto.response.CarsRacingDto;
+import racingcar.domain.car.dto.input.CreateCarsRacing;
+import racingcar.domain.car.dto.output.CarRacingDto;
+import racingcar.domain.car.dto.output.CarsRacingDto;
 import racingcar.domain.move.MoveCommand;
 import racingcar.domain.move.MoveCommander;
 import racingcar.exception.DuplicateCarNameException;
