+++ src/test/java/racingcar/domain/car/dto/CreateCarsTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.car.dto.request.CreateCarsRacing;
+import racingcar.domain.car.dto.input.CreateCarsRacing;
 import racingcar.exception.DuplicateCarNameException;
 import racingcar.exception.NoCarsException;
